fix: harden Kafka transport edge cases
Fixes:
- Flush the underlying Kafka producer from the transport producer flush path
- Preserve null package keys through header and legacy serialization modes
- Stop header splitting after returning an unsplittable oversized message
- Preserve header split segment counts above byte range while merging

Tests:
- Add regression coverage for transport flush, null keys, unsplittable messages, and 256 segment merges

+++ src/QuixStreams.Kafka.Transport/SerDes/KafkaMessageBuffer.cs
@@ -289,7 +289,7 @@ namespace QuixStreams.Kafka.Transport.SerDes
                 }
 
                 segmentLengths = buffer.MessageLength;
-                messageCount = (byte) buffer.MessageBuffer.Length;
+                messageCount = buffer.MessageBuffer.Length;
 
                 var val = buffer.MessageBuffer;
                 return val;
